doubles

+++ Down/Down/DivideCommand.cpp
@@ -14,8 +14,8 @@ void DivideCommand::execute(VirtualMachine& vm, AbstractFunctionCall& node)
 
 	if (variable1.getTokenType() == Token::TYPE_NUMBER && variable2.getTokenType() == Token::TYPE_NUMBER) {
 
-		int number1 = atoi(variable1.getValue().c_str());
-		int number2 = atoi(variable2.getValue().c_str());
+		double number1 = atof(variable1.getValue().c_str());
+		double number2 = atof(variable2.getValue().c_str());
 
 		if (number2 != 0) {
 			vm.setReturnValue(to_string(number1 / number2));
